Added handlers for Screening and Movie
Briefly tested, probably has a lot of errors.

+++ movietown/model/movie_type.go
@@ -3,5 +3,5 @@ package model
 type MovieType struct {
 	ID    uint `gorm:"primaryKey"`
 	Type  string
-	Price uint
+	Price float64
 }
